requirements.txt
torch
ninja
filelock
Pillow
torch
fire
humanize
requests
tqdm
matplotlib
scikit-image
scipy
numpy
blobfile
pyaml
clip @ git+https://github.com/openai/CLIP.git
